Add support for locations participating in NJVSS
This adds support for locations that are participating in the NJ state system. (I'm not sure who to mark as the "operator" in this case, but oh well.) This pulls from the same API that powers the map at https://covidvaccine.nj.gov/covidvaccine-location-map/. It appears as though it only returns results for locations that have availability, so that's how I'm currently treating it. BUT we might learn differently.

This is also clearly losing some data from the API. The data model for *everything* needs fixing, and we'll get to that shortly.