Add data from official locations list (#2)
This adds a CSV that mirrors all the data available in the state's official list of known vaccination locations (https://covid19.nj.gov/pages/covid-19-vaccine-locations-for-eligible-recipients). We may wish to see about live updating it either by scraping the site or by using the Google Sheets API.

The data here is used to fill in "no" values for providers that don't have a way to query each of their locations (usually because appointments are available at none of them).

**You can also specify `--include-all` on the CLI** to use this data to create "unknown" listings for all the locations that weren't scraped and aren't in the input data.

Matching things up between this data and the scraped sources is a bit messy, and I've written a few helper functions here to do relatively loose matches for cases where punctuation, capitalization, abbreviations, and so on differ. So far I've only noticed one place that I had to special case a manual match.