Auto copy `index.html` to `gh-pages` on push (#3)
Keep the source code for the "site" that renders data files with the rest of the code and copy it to the `gh-pages` branch when it's updated. We were previously keeping the canonical source on `gh-pages`, but this is much better for reviews and change management.

This also updates `index.html` with all kinds of fancy new new controls and layout bits. You can group and filter by county as well as search. It's definitely getting more complex than I'd ever really intended it to be, and it should probably flip over to something more framework-y with a real build process if we take it farther.

Finally, the `site` directory has a `data.json` file that serves as example scraper data to render `index.html` with. It's not an actual part of the site, though.